feat : output only thread 적용
